feat : CoverImage 테스트 코드 추가

+++ src/test/java/nextstep/courses/domain/CoverImageTest.java
@@ -15,8 +15,8 @@ public class CoverImageTest {
       // given
       String fileName = "filetest.png";
       long size = 10;
-    double width = 300;
-    double height = 200;
+    long width = 300;
+    long height = 200;
 
       // when
       CoverImage coverImage = CoverImage.defaultOf(fileName, size, width, height);
@@ -31,8 +31,8 @@ public class CoverImageTest {
     // given
     String fileName = "filetest.png";
     long size = 1024L*1024L;
-    double width = 300;
-    double height = 200;
+    long width = 300;
+    long height = 200;
 
     // then
     assertThrows(IllegalArgumentException.class
@@ -45,8 +45,8 @@ public class CoverImageTest {
     // given
     String fileName = "filetest.txt";
     long size = 1024L;
-    double width = 300;
-    double height = 200;
+    long width = 300;
+    long height = 200;
 
     // then
     assertThrows(IllegalArgumentException.class
@@ -59,8 +59,8 @@ public class CoverImageTest {
     // given
     String fileName = "filetest.png";
     long size = 1024L;
-    double width = 299;
-    double height = 200;
+    long width = 299;
+    long height = 200;
 
     // then
     assertThrows(IllegalArgumentException.class
@@ -73,8 +73,8 @@ public class CoverImageTest {
     // given
     String fileName = "filetest.png";
     long size = 1024L;
-    double width = 300;
-    double height = 199;
+    long width = 300;
+    long height = 199;
 
     // then
     assertThrows(IllegalArgumentException.class
@@ -87,8 +87,8 @@ public class CoverImageTest {
     // given
     String fileName = "filetest.png";
     long size = 1024L;
-    double width = 300;
-    double height = 201;
+    long width = 300;
+    long height = 201;
 
     // then
     assertThrows(IllegalArgumentException.class
